+++ src/app/directives/upload-age-directive/upload-age.directive.spec.ts
@@ -1,0 +1,8 @@
+import { UploadAgeDirective } from './upload-age.directive';
+
+describe('UploadAgeDirective', () => {
+  it('should create an instance', () => {
+    const directive = new UploadAgeDirective();
+    expect(directive).toBeTruthy();
+  });
+});
