+++ src/app/search-results/search-results.component.html
@@ -7,8 +7,13 @@
     @for(item of searchResults ; track item.id) {
     <mat-card
       appearance="outlined"
-      [ngClass]="getBorderColor(item.snippet.publishedAt)"
+      appUploadAge
+      [publishedAt]="item.snippet.publishedAt"
     >
+      <!-- <mat-card
+      appearance="outlined"
+      [ngClass]="getBorderColor(item.snippet.publishedAt)"
+    > -->
       <mat-card-header>
         <img
           mat-card-image
